add new res

+++ 3.8.1/Voxel_Explorer/Classes/Levels/FaneBossLevel.cpp
@@ -360,7 +360,7 @@ bool FaneBossLevel::createPickableItems()
         if(neighbourHasDoor)
             continue;
         
-        int playerlevel = PlayerProperty::getInstance()->getLevel().GetLongValue();
+        int playerlevel = (int)PlayerProperty::getInstance()->getLevel().GetLongValue();
         int itemLevel = cocos2d::random(playerlevel - 3, playerlevel + 3);
         if(itemLevel < 1)
             itemLevel = 1;
