+++ 3.8.1/Voxel_Explorer/Classes/Sprites3D/Boss/SkeletonKing.cpp
@@ -229,7 +229,7 @@ void SkeletonKing::onEnterDeath()
     if(m_pFakeShadow)
         m_pFakeShadow->setVisible(false);
     VoxelExplorer::getInstance()->addParticle3DEffect(getPosition3D(), P3D_EFFECT_TYPE::P3D_PLAYER_DEATH);
-    VoxelExplorer::getInstance()->generatePickItemByBoss(getPosInMap(), m_pBossProperty->getValueCopper().GetLongValue());
+    VoxelExplorer::getInstance()->generatePickItemByBoss(getPosInMap(), (int)m_pBossProperty->getValueCopper().GetLongValue());
     
     TombBossLevel* level = dynamic_cast<TombBossLevel*>(VoxelExplorer::getInstance()->getCurrentLevel());
     if(level)
